better ordering of module namespace
