+++ src/actions/list/mod.rs
@@ -3,13 +3,9 @@ extern crate serde_json;
 
 use self::curl::http;
 use std::str::from_utf8;
-use std::io::prelude::*;
-use std::fs::File;
-use mods::*;
+use actions::*;
 
-pub fn export(server: &str, port: &str, key: &str, file: &str, verbose: bool) {
-
-    let mut file = File::create(file).unwrap();
+pub fn list(server: &str, port: &str, key: &str, verbose: bool) {
 
     // build url from input values
     // format the url depending on if a key is set
@@ -49,7 +45,6 @@ pub fn export(server: &str, port: &str, key: &str, file: &str, verbose: bool) {
     // map json body to our backend Struct
     let json: Vec<ValueData> = serde_json::from_str(&body[..]).unwrap();
 
-    let output = decode_json(json).unwrap();
-
-    file.write_all(output.as_bytes()).unwrap()
+    // pass json to our decode_json function and print result
+    println!("{}", decode_json(json).unwrap());
 }
