+++ src/actions/remove/mod.rs
